fix: stabilize Ketua Tim KPI dashboard rendering (closes #30)
- Rewrite KetuaTimDashboard.tsx: fix Alpine.js lifecycle timing bug
- Root cause: x-data inline object broke HTML attribute parsing due to
  single quotes inside JS arrays conflicting with HTML attr delimiters
- Solution: use Alpine.data('kpiDash') registered via alpine:init listener
  with x-data double-quoted attribute — no quote conflicts
- Add credentials:'include' to fetch so auth cookie is sent to API
- Add kpiError state with visible error panel + retry button
- Add Chart.js 4.4.0 Spider Chart (Radar) with try/catch guard
- Add loading skeleton with pulse animation
- Separate KetuaTimAllProjects from dashboard (single responsibility)
- Route /dashboard/projects -> KetuaTimAllProjects (project management)
- Route /dashboard/ketua -> KetuaTimDashboard (KPI analytics only)
